fix(#7199): move the rollback tests into GlobalsTest so jtcop finds their production class
